[APDV-111] Moved group members mappers to separate package
