+++ backend/src/main/java/pl/edu/agh/apdvbackend/mappers/group/group_members/GroupMembersMapper.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.apdvbackend.mappers.group;
+package pl.edu.agh.apdvbackend.mappers.group.group_members;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
