+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/group/GetGroupMembersImpl.java
@@ -2,7 +2,7 @@ package pl.edu.agh.apdvbackend.use_cases.group;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.edu.agh.apdvbackend.mappers.group.GroupMembersMapper;
+import pl.edu.agh.apdvbackend.mappers.group.group_members.GroupMembersMapper;
 import pl.edu.agh.apdvbackend.models.body_models.group.GroupMembersResponseBody;
 
 @Component
